+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2023-01-25
+
 ### Added
 
 - `--effective` argument, to use Ledger's effective dates